+++ src/extract/mod.rs
@@ -14,6 +14,7 @@ pub mod aspack;
 pub mod bcm;
 pub mod chdman;
 pub mod cic;
+pub mod daa;
 pub mod dispatch;
 pub mod extsis;
 pub mod freearc;